requirements.txt
Django==5.0.6
djangorestframework==3.15.2
gunicorn==22.0.0
python-dotenv==1.0.1
psycopg[binary]==3.1.19
httpx==0.27.0
beautifulsoup4==4.12.3
python-dateutil==2.9.0.post0
whitenoise==6.7.0
